+++ config/initializers/categories.rb
@@ -1,0 +1,4 @@
+hashes = YAML.load_file(Rails.root.join('config/categories.yml'))
+STATIC_CATEGORIES = hashes.collect do |hash|
+  OpenStruct.new hash
+end